Update angular and start refactoring process JC-645
