+++ apps/angular/src/theme/variables.css
@@ -7,9 +7,10 @@
   --on-accent: rgb(14 8 48 / 100%);
 
   /* Space variables */
-  --space-5: 20px;
-  --space-6: 24px;
-  --space-8: 32px;
-  --space-10: 40px;
-  --space-20: 80px;
+  --space-base: 4px;
+  --space-5: calc(5 * var(--space-base));
+  --space-6: calc(6 * var(--space-base));
+  --space-8: calc(8 * var(--space-base));
+  --space-10: calc(10 * var(--space-base));
+  --space-20: calc(20 * var(--space-base));
 }
