+++ libs/core/dtos/time-interval.dto.ts
@@ -1,5 +1,5 @@
 /** Anime show dates DTO. */
-export type AiredDto = {
+export type TimeIntervalDto = {
 
 	/** Start airing date in format 2024-07-22T02:40:54.873Z. */
 	readonly start: string | null;
