+++ libs/core/utils/enums/dto-status.enum.ts
@@ -4,10 +4,3 @@ export enum DtoStatus {
 	Finished = 'FINISHED',
 	NotYetAired = 'NOT_YET_AIRED',
 }
-
-/** Represents anime status in the model. */
-export enum ModelStatus{
-	Airing = 'airing',
-	Finished = 'finished',
-	NotYetAired = 'expected',
-}
