+++ libs/core/utils/enums/dto-type.enum.ts
@@ -9,15 +9,3 @@ export enum DtoType {
 	PromotionalVideos = 'PROMOTIONAL_VIDEOS',
 	Unknown = 'UNKNOWN',
 }
-
-/** Represents anime type in the model. */
-export enum ModelType{
-	TV = 'TV',
-	OVA = 'OVA',
-	Movie = 'movie',
-	Special = 'special',
-	ONA = 'ONA',
-	Music = 'music',
-	PromotionalVideos = 'promotional videos',
-	Unknown = '',
-}
