+++ libs/core/utils/enums/model-status.enum.ts
@@ -1,0 +1,6 @@
+/** Represents anime status in the model. */
+export enum ModelStatus{
+	Airing = 'airing',
+	Finished = 'finished',
+	NotYetAired = 'expected',
+}
