+++ libs/core/utils/enums/model-type.enum.ts
@@ -1,0 +1,11 @@
+/** Represents anime type in the model. */
+export enum ModelType{
+	TV = 'TV',
+	OVA = 'OVA',
+	Movie = 'movie',
+	Special = 'special',
+	ONA = 'ONA',
+	Music = 'music',
+	PromotionalVideos = 'promotional videos',
+	Unknown = '',
+}
