+++ package.json
@@ -38,6 +38,7 @@
     "@angular/platform-browser-dynamic": "18.0.5",
     "@angular/router": "18.0.5",
     "@reduxjs/toolkit": "1.9.3",
+    "angular": "^1.8.3",
     "axios": "^1.4.0",
     "core-js": "^3.30.2",
     "react": "18.3.1",
